+++ src/Exchange.js
@@ -71,7 +71,7 @@ function ExchangeBase(name, isExchange, type = 'topic', options = {}, eventExcha
 
   function shouldIPublish(messageProperties) {
     if (stopped) return;
-    if (messageProperties.mandatory) return true;
+    if (messageProperties.mandatory || messageProperties.confirm) return true;
     return bindings.length;
   }
 
@@ -81,9 +81,7 @@ function ExchangeBase(name, isExchange, type = 'topic', options = {}, eventExcha
 
     if (!deliverTo.length) {
       message.ack();
-      if (publishedMsg.properties.mandatory) {
-        emitReturn(routingKey, publishedMsg);
-      }
+      emitReturn(routingKey, publishedMsg);
       return 0;
     }
 
@@ -98,9 +96,7 @@ function ExchangeBase(name, isExchange, type = 'topic', options = {}, eventExcha
     const first = deliverTo[0];
     if (!first) {
       message.ack();
-      if (publishedMsg.properties.mandatory) {
-        emitReturn(routingKey, publishedMsg);
-      }
+      emitReturn(routingKey, publishedMsg);
       return 0;
     }
 
@@ -115,7 +111,13 @@ function ExchangeBase(name, isExchange, type = 'topic', options = {}, eventExcha
   }
 
   function emitReturn(routingKey, returnMessage) {
-    emit('return', Message({routingKey, exchange: name}, returnMessage.content, returnMessage.properties));
+    const {content, properties} = returnMessage;
+    if (properties.confirm) {
+      emit('message.undelivered', Message({routingKey, exchange: name}, content, properties));
+    }
+    if (properties.mandatory) {
+      emit('return', Message({routingKey, exchange: name}, content, properties));
+    }
   }
 
   function getConcernedBindings(routingKey) {
@@ -223,7 +225,7 @@ function ExchangeBase(name, isExchange, type = 'topic', options = {}, eventExcha
   }
 
   function on(pattern, handler, consumeOptions = {}) {
-    if (isExchange) return eventExchange.on(`exchange.${pattern}`, handler);
+    if (isExchange) return eventExchange.on(`exchange.${pattern}`, handler, consumeOptions);
 
     const eventQueue = Queue(null, {durable: false, autoDelete: true});
     bind(eventQueue, pattern);
@@ -234,8 +236,10 @@ function ExchangeBase(name, isExchange, type = 'topic', options = {}, eventExcha
   function off(pattern, handler) {
     if (isExchange) return eventExchange.off(`exchange.${pattern}`, handler);
 
+    const {consumerTag} = handler;
     for (const binding of bindings) {
       if (binding.pattern === pattern) {
+        if (consumerTag) binding.queue.cancel(consumerTag);
         binding.queue.dismiss(handler);
       }
     }